requirements.txt
python-dotenv==0.21.0
requests==2.28.1
rumps==0.4.0